Moved chip defintion into chip.rs so that other modules can name and use it. This is so, for example, an interrupt handler can insert an entry into the interrupt deferred call queue.
